Update wording for each notification setting
