+++ pages/settings.vue
@@ -9,7 +9,7 @@
                         <nuxt-link to="profile">Profile</nuxt-link>
                         <nuxt-link to="account">Account</nuxt-link>
                         <nuxt-link to="connections">Connections</nuxt-link>
-                        <nuxt-link to="email">Email Settings</nuxt-link>
+                        <nuxt-link to="notifications">Notifications</nuxt-link>
                     </Tabs>
 
                     <Card class="plain dark">
@@ -28,10 +28,12 @@ import PageBanner from '@/components/layout/PageBanner';
 import { names } from '../utils/auth';
 
 export default {
-    name: 'Settings',
+    name: 'SettingsPage',
+
     meta: {
         auth: names.USER,
     },
+
     components: { Tabs, Card, PageBanner },
 };
 </script>
